Refactored code into multiple smaller files, made it more readable
