+++ Position.h
@@ -1,0 +1,9 @@
+#ifndef POSITION_H
+#define POSITION_H
+
+typedef struct {
+    int row;
+    int col;
+} Position;
+
+#endif // POSITION_H
