--- game_history.txt
@@ -1,36 +1,0 @@
-rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w - - 0 0
-rnbqkbnr/pppppppp/8/8/8/5N2/PPPPPPPP/RNBQKB1R b - - 1 0
-rnbqkb1r/pppppppp/5n2/8/8/5N2/PPPPPPPP/RNBQKB1R w - - 2 1
-rnbqkb1r/pppppppp/5n2/4N3/8/8/PPPPPPPP/RNBQKB1R b - - 3 1
-rnbqkb1r/pppppppp/8/4N3/4n3/8/PPPPPPPP/RNBQKB1R w - - 4 2
-rnbqkb1r/pppppppp/6N1/8/4n3/8/PPPPPPPP/RNBQKB1R b - - 5 2
-rnbqkb1r/pppppppp/6N1/8/8/6n1/PPPPPPPP/RNBQKB1R w - - 6 3
-rnbqkb1r/pppppppp/8/8/7N/6n1/PPPPPPPP/RNBQKB1R b - - 7 3
-rnbqkb1r/pppppppp/8/7n/7N/8/PPPPPPPP/RNBQKB1R w - - 8 4
-rnbqkb1r/pppppppp/8/7n/7N/2N5/PPPPPPPP/R1BQKB1R b - - 9 4
-r1bqkb1r/pppppppp/2n5/7n/7N/2N5/PPPPPPPP/R1BQKB1R w - - 10 5
-r1bqkb1r/pppppppp/2n5/7n/8/2N2N2/PPPPPPPP/R1BQKB1R b - - 11 5
-r1bqkb1r/pppppppp/2n2n2/8/8/2N2N2/PPPPPPPP/R1BQKB1R w - - 12 6
-r1bqkb1r/pppppppp/2n2n2/8/3N4/2N5/PPPPPPPP/R1BQKB1R b - - 13 6
-r1bqkb1r/pppppppp/2n5/3n4/3N4/2N5/PPPPPPPP/R1BQKB1R w - - 14 7
-r1bqkb1r/pppppppp/2n1N3/3n4/8/2N5/PPPPPPPP/R1BQKB1R b - - 15 7
-r1bqkb1r/pppppppp/2n1N3/8/8/2N1n3/PPPPPPPP/R1BQKB1R w - - 16 8
-r1bqkb1r/pppppppp/2n5/6N1/8/2N1n3/PPPPPPPP/R1BQKB1R b - - 17 8
-r1bqkb1r/pppppppp/2n5/6N1/6n1/2N5/PPPPPPPP/R1BQKB1R w - - 18 9
-r1bqkb1r/pppppppp/2n5/8/6n1/2N4N/PPPPPPPP/R1BQKB1R b - - 19 9
-r1bqkb1r/pppppppp/2n4n/8/8/2N4N/PPPPPPPP/R1BQKB1R w - - 20 10
-r1bqkb1r/pppppppp/2n4n/6N1/8/2N5/PPPPPPPP/R1BQKB1R b - - 21 10
-r1bqkb1r/pppppppp/2n5/6N1/6n1/2N5/PPPPPPPP/R1BQKB1R w - - 22 11
-r1bqkb1r/pppppppp/2n1N3/8/6n1/2N5/PPPPPPPP/R1BQKB1R b - - 23 11
-r1bqkb1r/pppppppp/2n1N3/8/8/2N1n3/PPPPPPPP/R1BQKB1R w - - 24 12
-r1bqkb1r/pppppppp/2n5/2N5/8/2N1n3/PPPPPPPP/R1BQKB1R b - - 25 12
-r1bqkb1r/pppppppp/2n5/2N5/2n5/2N5/PPPPPPPP/R1BQKB1R w - - 26 13
-r1bqkb1r/pppppppp/N1n5/8/2n5/2N5/PPPPPPPP/R1BQKB1R b - - 27 13
-r1bqkb1r/pppppppp/N1n5/8/8/n1N5/PPPPPPPP/R1BQKB1R w - - 28 14
-r1bqkb1r/pppppppp/2n5/8/1N6/n1N5/PPPPPPPP/R1BQKB1R b - - 29 14
-r1bqkb1r/pppppppp/2n5/1n6/1N6/2N5/PPPPPPPP/R1BQKB1R w - - 30 15
-r1bqkb1r/pppppppp/2n5/1n1N4/8/2N5/PPPPPPPP/R1BQKB1R b - - 31 15
-r1bqkb1r/pppppppp/2n5/3N4/3n4/2N5/PPPPPPPP/R1BQKB1R w - - 32 16
-r1bqkb1r/pppppppp/2n2N2/8/3n4/2N5/PPPPPPPP/R1BQKB1R b - - 33 16
-r1bqkb1r/pppp1ppp/2n2p2/8/3n4/2N5/PPPPPPPP/R1BQKB1R w - - 34 17
-r1bqkb1r/pppp1ppp/2n2p2/3N4/3n4/8/PPPPPPPP/R1BQKB1R b - - 35 17
